backend/pyproject.toml
[tool.poetry]
name = "bookstore-api"
version = "0.1.0"
description = ""
repository = ""
authors = ["Hussein Auf <hu.auf2000@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.11"
django-environ = "^0.9.0"
celery = "^5.2.7"
psycopg2 = "^2.9.3"
redis = "^4.3.4"
django-celery-beat = "^2.3.0"
django-celery-results = "^2.4.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
django-cors-headers = "^3.13.0"
django-model-utils = "^5.0.0"
django-filter = "^24.3"
drf-nested-routers = "^0.94.1"
pillow = "^10.4.0"
django-debug-toolbar = "^4.4.6"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
isort = "^5.13.2"
flake8 = "^7.1.1"
django-extensions = "^3.2.3"
drf-spectacular = "^0.27.2"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}

[tool.poetry.group.dev.dependencies]
pylint = "2.15.3"
black = "22.8.0"
pylint-django = "^2.5.3"
pylint-celery = "^0.3"
coverage = "^6.5.0"
safety = "^2.2.0"
django-coverage-plugin = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
ensure_newline_before_comments = false
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
  | \.git
  | \.mypy_cache
  | \.venv
  | dist
  | migrations
)/
'''

[tool.isort]
profile = 'black'
